# api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErroResposta": {
        "properties": {
          "errors": {
            "items": {
              "properties": {
                "campo": {
                  "type": "string"
                },
                "mensagem": {
                  "type": "string"
                }
              },
              "required": [
                "campo",
                "mensagem"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          },
          "statusCode": {
            "maximum": 9007199254740991,
            "minimum": -9007199254740991,
            "type": "integer"
          }
        },
        "required": [
          "statusCode",
          "message"
        ],
        "type": "object"
      },
      "Login": {
        "properties": {
          "email": {
            "type": "string"
          },
          "senha": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "senha"
        ],
        "type": "object"
      },
...
      "RejeicaoEmpresa": {
        "properties": {
          "motivo": {
            "type": "string"
          }
        },
        "required": [
          "motivo"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "cookie": {
        "in": "cookie",
        "name": "fivo_sessao",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "API de cadastro e autenticação de empresas",
    "title": "Fivo — Selo Isso Importa",
    "version": "0.0.1"
  },
  "openapi": "3.0.0",
  "paths": {
    "/admin/empresas": {
      "get": {
        "operationId": "AdminEmpresasController_listar",
        "parameters": [
          {
            "in": "query",
            "name": "estado",
            "required": false,
            "schema": {
              "enum": [
                "PENDENTE_APROVACAO"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "properties": {
                      "cnpj": {
                        "type": "string"
                      },
                      "criadoEm": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "id": {
                        "type": "string"
                      },
                      "razaoSocial": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "razaoSocial",
                      "cnpj",
                      "email",
                      "criadoEm"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Empresas pendentes, da mais antiga para a mais recente"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Sessão ausente, inválida ou expirada"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Acesso negado"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Estado não suportado"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Fila de empresas pendentes de aprovação (somente ADMIN)",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/empresas/{id}/aprovacao": {
      "post": {
        "operationId": "AdminEmpresasController_aprovar",
        "parameters": [
          {
            "description": "Id da empresa",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Decisão registrada"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Sessão ausente, inválida ou expirada"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Acesso negado"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Empresa não encontrada"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Transição de estado inválida para a situação atual"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Aprova a empresa (somente ADMIN)",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/empresas/{id}/reativacao": {
      "post": {
        "operationId": "AdminEmpresasController_reativar",
        "parameters": [
          {
            "description": "Id da empresa",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Decisão registrada"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Sessão ausente, inválida ou expirada"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Acesso negado"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Empresa não encontrada"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Transição de estado inválida para a situação atual"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Reativa a empresa (somente ADMIN)",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/empresas/{id}/rejeicao": {
      "post": {
        "operationId": "AdminEmpresasController_rejeitar",
        "parameters": [
          {
            "description": "Id da empresa",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejeicaoEmpresa"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Decisão registrada"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Sessão ausente, inválida ou expirada"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Acesso negado"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Empresa não encontrada"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Transição de estado inválida para a situação atual"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Motivo ausente ou insuficiente"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Rejeita a empresa com motivo (somente ADMIN)",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/empresas/{id}/suspensao": {
      "post": {
        "operationId": "AdminEmpresasController_suspender",
        "parameters": [
          {
            "description": "Id da empresa",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Decisão registrada"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Sessão ausente, inválida ou expirada"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Acesso negado"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Empresa não encontrada"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Transição de estado inválida para a situação atual"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Suspende a empresa (somente ADMIN)",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/sessoes": {
      "post": {
        "operationId": "AutenticacaoController_entrar",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "papel": {
                      "enum": [
                        "ADMIN",
                        "EMPRESA",
                        "INSTITUICAO"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "papel"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Sessão aberta",
            "headers": {
              "Set-Cookie": {
                "description": "Cookie de sessão opaca (HttpOnly)",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Credenciais inválidas"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Dados inválidos"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Conta bloqueada por excesso de tentativas"
          }
        },
        "summary": "Abre a sessão (login) e define o cookie",
        "tags": [
          "sessoes"
        ]
      }
    },
    "/sessoes/atual": {
      "delete": {
        "operationId": "AutenticacaoController_sair",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Sessão encerrada"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            },
            "description": "Sessão ausente, inválida ou expirada"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Encerra a sessão atual (logout)",
        "tags": [
          "sessoes"
        ]
      }
    }
  },
  "servers": [],
...
}